do favorite

+++ apps/cart/views.py
@@ -40,27 +40,27 @@ class CartItemListSet(viewsets.ModelViewSet):
 class FavoriteCreateSet(viewsets.ModelViewSet):
     queryset = FavoriteProduct.objects.all()
     serializer_class = FavoriteSerializer
-    permission_classes = [IsAuthenticated, ]
+    lookup_field = 'id'
 
     def destroy(self, request, *args, **kwargs):
         if 'pk' in kwargs:
-            try:
-                cart_item = self.get_queryset().get(pk=kwargs['pk'])
+            cart_item = self.get_object()
+            if cart_item:
                 cart_item.delete()
                 return Response(
-                    {"message": " Избранный товар был успешно удален из корзины."},
+                    {"message": "Товар был успешно удален из корзины."},
                     status=status.HTTP_204_NO_CONTENT
                 )
-            except FavoriteProduct.DoesNotExist:
+            else:
                 return Response(
-                    {"message": "Избранный товар не найден в корзине."},
+                    {"message": "Товар не найден в корзине."},
                     status=status.HTTP_404_NOT_FOUND
                 )
         else:
-            favorite_items = self.get_queryset()
-            favorite_items.delete()
+            cart_items = self.get_queryset()
+            cart_items.delete()
             return Response(
-                {"message": "Избранные были успешно очищена."},
+                {"message": "Корзина была успешно очищена."},
                 status=status.HTTP_200_OK
             )
